Moved location-related seed classes to the new directory
